+++ src/main/java/shadows/apotheosis/adventure/affix/effect/ExecutingAffix.java
@@ -5,15 +5,19 @@ import java.util.function.Consumer;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.EntityDamageSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
+import shadows.apotheosis.Apotheosis;
 import shadows.apotheosis.adventure.affix.Affix;
 import shadows.apotheosis.adventure.affix.AffixHelper;
 import shadows.apotheosis.adventure.affix.AffixType;
 import shadows.apotheosis.adventure.loot.LootCategory;
 import shadows.apotheosis.adventure.loot.LootRarity;
+import shadows.apotheosis.mixin.LivingEntityInvoker;
 import shadows.placebo.util.StepFunction;
 
 public class ExecutingAffix extends Affix {
@@ -41,9 +45,18 @@ public class ExecutingAffix extends Affix {
 	@Override
 	public void doPostAttack(ItemStack stack, LootRarity rarity, float level, LivingEntity user, Entity target) {
 		float threshold = getTrueLevel(rarity, level);
-		if (target instanceof LivingEntity living && !living.level.isClientSide) {
+		if (Apotheosis.localAtkStrength >= 0.98 && target instanceof LivingEntity living && !living.level.isClientSide) {
 			if (living.getHealth() / living.getMaxHealth() < threshold) {
-				living.die(new EntityDamageSource("apotheosis.execute", user).bypassInvul().bypassArmor().bypassMagic());
+				DamageSource src = new EntityDamageSource("apotheosis.execute", user).bypassArmor().bypassMagic();
+				if (!((LivingEntityInvoker) living).callCheckTotemDeathProtection(src)) {
+					SoundEvent soundevent = ((LivingEntityInvoker) living).callGetDeathSound();
+					if (soundevent != null) {
+						living.playSound(soundevent, ((LivingEntityInvoker) living).callGetSoundVolume(), living.getVoicePitch());
+					}
+
+					living.setHealth(0);
+					living.die(src);
+				}
 			}
 		}
 	}