+++ src/main/java/shadows/apotheosis/adventure/compat/AdventureJEIPlugin.java
@@ -10,8 +10,10 @@ import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier.Operation;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.registries.IRegistryDelegate;
 import shadows.apotheosis.Apoth;
 import shadows.apotheosis.Apotheosis;
+import shadows.apotheosis.adventure.AdventureModule;
 import shadows.apotheosis.adventure.affix.socket.GemItem;
 
 @JeiPlugin
@@ -31,6 +33,7 @@ public class AdventureJEIPlugin implements IModPlugin {
 		reg.addIngredientInfo(new ItemStack(Apoth.Items.GEM_DUST), VanillaTypes.ITEM, new TranslatableComponent("info.apotheosis.gem_crushing"));
 		reg.addIngredientInfo(new ItemStack(Apoth.Items.VIAL_OF_EXTRACTION), VanillaTypes.ITEM, new TranslatableComponent("info.apotheosis.gem_extraction"));
 		reg.addIngredientInfo(new ItemStack(Apoth.Items.VIAL_OF_EXPULSION), VanillaTypes.ITEM, new TranslatableComponent("info.apotheosis.gem_expulsion"));
+		reg.addIngredientInfo(AdventureModule.RARITY_MATERIALS.values().stream().map(IRegistryDelegate::get).map(ItemStack::new).toList(), VanillaTypes.ITEM, new TranslatableComponent("info.apotheosis.salvaging"));
 	}
 
 }