+++ src/main/java/shadows/apotheosis/mixin/LivingEntityInvoker.java
@@ -3,6 +3,7 @@ package shadows.apotheosis.mixin;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
+import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
 
@@ -14,4 +15,16 @@ public interface LivingEntityInvoker {
 	 */
 	@Invoker
 	public void callActuallyHurt(DamageSource pDamageSrc, float pDamageAmount);
+
+	/**
+	 * Invokes the private method {@link LivingEntity#checkTotemDeathProtection(DamageSource)}
+	 */
+	@Invoker
+	public boolean callCheckTotemDeathProtection(DamageSource pDamageSource);
+
+	@Invoker
+	public SoundEvent callGetDeathSound();
+
+	@Invoker
+	public float callGetSoundVolume();
 }